Fixed rendering test in UIScrollingContainer tests. Renamed 'comparing functions to comparators
