+++ tests/test_elements/test_ui_scrolling_container.py
@@ -4,7 +4,7 @@ import pygame
 
 from tests.shared_fixtures import _init_pygame, default_ui_manager
 from tests.shared_fixtures import default_display_surface , _display_surface_return_none
-from tests.comparing_functions import compare_surfaces
+from tests.comparators import compare_surfaces
 
 from pygame_gui.ui_manager import UIManager
 from pygame_gui.elements.ui_scrolling_container import UIScrollingContainer
@@ -250,10 +250,10 @@ class TestUIScrollingContainer:
         surface = empty_surface.copy()
         manager = UIManager(resolution)
 
-        container = UIScrollingContainer(relative_rect=pygame.Rect(100, 100, 400, 400),
+        container = UIScrollingContainer(relative_rect=pygame.Rect(100, 100, 200, 100),
                                          manager=manager,
                                          visible=0)
-        container.set_scrollable_area_dimensions((500, 600))
+        container.set_scrollable_area_dimensions((600, 600))
 
         manager.draw_ui(surface)
         assert compare_surfaces(empty_surface, surface)
