+++ tests/test_elements/test_ui_vertical_scroll_bar.py
@@ -4,7 +4,7 @@ import pygame
 
 from tests.shared_fixtures import _init_pygame, default_ui_manager
 from tests.shared_fixtures import default_display_surface , _display_surface_return_none
-from tests.comparing_functions import compare_surfaces
+from tests.comparators import compare_surfaces
 
 from pygame_gui.ui_manager import UIManager
 from pygame_gui.elements.ui_vertical_scroll_bar import UIVerticalScrollBar
