+++ tests/test_elements/test_ui_window.py
@@ -5,7 +5,7 @@ import pygame_gui
 
 from tests.shared_fixtures import _init_pygame, default_ui_manager
 from tests.shared_fixtures import default_display_surface, _display_surface_return_none
-from tests.comparing_functions import compare_surfaces
+from tests.comparators import compare_surfaces
 
 from pygame_gui.elements.ui_window import UIWindow
 from pygame_gui.elements.ui_button import UIButton
